backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
httpx>=0.27.0
redis>=5.2.0
psycopg2-binary>=2.9.10
alembic>=1.14.0
